Removed VOLUME instruction
`VOLUME` instruction should not be used anymore because it creates a new volume on each container creation. If you remove the container the volume will be orphan and the data inaccessible (by containers). As a result, we have a lot of disk space used with old data.

The docker recommendation is to use `host volumes` or `named volumes` (recommended).